ref: rename module with assertions (#88)
