+++ lib/e2e/src/lib.rs
@@ -1,11 +1,11 @@
-mod assert;
+mod assertions;
 mod deploy;
 mod environment;
 mod project;
 mod system;
 mod user;
 
-pub use assert::{Assert, Emits};
+pub use assertions::{Assert, Emits};
 pub use deploy::deploy;
 pub use e2e_proc::test;
 pub use system::{provider, Provider, Signer};
